TEIID-3189 column datatypes were changed and requires the create table statement to change to datatype of long

+++ dynamicvdb-datafederation/src/teiidfiles/customer-schema.sql
@@ -17,7 +17,7 @@ CREATE TABLE status
   TargetName varchar(256) not null,
   Valid boolean not null,
   LoadState varchar(25) not null,
-  Cardinality integer,
+  Cardinality long,
   Updated timestamp not null,
   LoadNumber long not null,
   PRIMARY KEY (VDBName, VDBVersion, SchemaName, Name)
